requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
black==24.10.0
click==8.1.8
Django==5.1.1
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
